Added support for a "token" query parameter.
